[Monitoring] Convert Kibana-related server files that read from _source to typescript (#86364)
* Identifying kibana source fields

* Fix types

* Fix types, take 2

* Should really fix it now
