+++ x-pack/plugins/monitoring/server/plugin.ts
@@ -335,6 +335,7 @@ export class Plugin {
               }
             },
             server: {
+              route: () => {},
               config: legacyConfigWrapper,
               newPlatform: {
                 setup: {
